Updating unit tests
