+++ gcp_icone_translator.py
@@ -1,13 +1,13 @@
-from translator import icone_translator
-from translator.tools import wzdx_translator
 import json
-import xmltodict
-import urllib
-from google.cloud import pubsub_v1
-import os
-from google.cloud import secretmanager
 import logging
-import jsonschema
+import os
+import urllib
+
+import xmltodict
+from google.cloud import secretmanager
+
+from translator import icone_translator
+from translator.tools import gcp_tools, wzdx_translator
 
 
 def get_ftp_url():
@@ -38,20 +38,18 @@ def main(event, context):
     icone_obj = parse_xml(icone_data)
 
     wzdx_obj = icone_translator.wzdx_creator(
-        icone_obj, unsupported_message_callback=unsupported_messages_callback)
-    wzdx_schema = get_wzdx_schema('wzdx_schema.json')
+        icone_obj, unsupported_message_callback=gcp_tools.unsupported_messages_callback)
+
+    wzdx_schema = gcp_tools.get_wzdx_schema(
+        'translator/sample files/validation_schema/wzdx_v3.1_feed.json')
     if not wzdx_translator.validate_wzdx(wzdx_obj, wzdx_schema):
         logging.error(RuntimeError(
             'WZDx message failed validation. Exiting Application !'))
         return 'WZDx message failed validation. Exiting Application !', 500
 
-    publisher = pubsub_v1.PublisherClient()
-    topic_path = publisher.topic_path(
-        os.environ['project_id'], os.environ['wzdx_topic_id'])
-    future = publisher.publish(topic_path, str.encode(json.dumps(
-        wzdx_obj, indent=2)), origin='auto_icone_translator_ftp cloud function')
-    logging.debug(future.result())
-    return
+    gcp_tools.publish_wzdx_message(wzdx_obj)
+
+    return "Success"
 
 
 def get_ftp_file(url):
@@ -62,27 +60,7 @@ def parse_xml(xml_string):
     return xmltodict.parse(xml_string)
 
 
-def get_wzdx_schema(schema_file_name):
-    try:
-        schema_string = open(schema_file_name, 'r').read()
-    except FileNotFoundError as e:
-        raise RuntimeError('invalid schema: file does not exist')
-    if not schema_string:
-        raise RuntimeError('invalid schema: empty schema')
-    try:
-        schema_obj = json.loads(schema_string)
-        jsonschema.validate('', schema_obj)
-    except json.decoder.JSONDecodeError as e:
-        raise RuntimeError('invalid schema: not valid json') from e
-    except jsonschema.SchemaError as e:
-        raise RuntimeError('invalid schema: schema failed validation') from e
-    except jsonschema.ValidationError:
-        return schema_obj
-    return schema_obj
-
-
 def get_ftp_credentials():
-
     secret_client = secretmanager.SecretManagerServiceClient()
     username_secret_name = os.environ.get('icone_ftp_username_secret_name')
     password_secret_name = os.environ.get('icone_ftp_password_secret_name')
@@ -111,33 +89,3 @@ def get_ftp_credentials():
     password = response.payload.data.decode("UTF-8")
 
     return username, password
-
-
-def unsupported_messages_callback(message):
-    project_id = os.environ.get('project_id')
-    unsupported_messages_topic_id = os.environ.get(
-        'unsupported_messages_topic_id')
-    if not project_id or not unsupported_messages_topic_id:
-        return False
-    publisher = pubsub_v1.PublisherClient()
-    topic_path = publisher.topic_path(
-        project_id, unsupported_messages_topic_id)
-    # publish unsupported messages into pub/sub topic
-    try:
-        future = publisher.publish(topic_path, str.encode(formatMessage(
-            message)), origin='auto_icone_translator_ftp cloud function')
-    except Exception as e:
-        logging.error('failed to publish unsupported message to project_id: {:s}, topic_id: {:s}, error_message: {:s}'.format(
-            project_id, unsupported_messages_topic_id, str(e)))
-        return False
-    logging.debug(future.result())
-    return True
-
-
-def formatMessage(message) -> str:
-    message_type = [dict, list, tuple, str, int, float, bool, type(None)]
-    if type(message) in message_type:
-        msg = json.dumps(message, indent=2)
-    else:
-        msg = str(message)
-    return msg
